pipeline: rename pipilines to exporters
