+++ Makefile
@@ -436,25 +436,25 @@ static: skydive.clean govendor genlocalfiles
 	$(MAKE) compile.static WITH_LIBVIRT_GO=false
 
 .PHONY: contribs.clean
-contribs.clean: contrib.pipelines.clean contrib.snort.clean contrib.collectd.clean
+contribs.clean: contrib.exporters.clean contrib.snort.clean contrib.collectd.clean
 
 .PHONY: contribs.test
-contribs.test: contrib.pipelines.test
+contribs.test: contrib.exporters.test
 
 .PHONY: contribs
-contribs: contrib.pipelines contrib.snort contrib.collectd
+contribs: contrib.exporters contrib.snort contrib.collectd
 
-.PHONY: contrib.pipelines.clean
-contrib.pipelines.clean:
-	$(MAKE) -C contrib/pipelines clean
+.PHONY: contrib.exporters.clean
+contrib.exporters.clean:
+	$(MAKE) -C contrib/exporters clean
 
-.PHONY: contrib.pipelines
-contrib.pipelines: govendor genlocalfiles
-	$(MAKE) -C contrib/pipelines
+.PHONY: contrib.exporters
+contrib.exporters: govendor genlocalfiles
+	$(MAKE) -C contrib/exporters
 
-.PHONY: contrib.pipelines.test
-contrib.pipelines.test: govendor genlocalfiles
-	$(MAKE) -C contrib/pipelines test
+.PHONY: contrib.exporters.test
+contrib.exporters.test: govendor genlocalfiles
+	$(MAKE) -C contrib/exporters test
 
 .PHONY: contribs.snort.clean
 contrib.snort.clean:
