+++ contrib/exporters/awsflowlogs/main.go
@@ -18,16 +18,14 @@
 package main
 
 import (
-	aws "github.com/skydive-project/skydive/contrib/pipelines/awsflowlogs/mod"
-	"github.com/skydive-project/skydive/contrib/pipelines/core"
-	sa "github.com/skydive-project/skydive/contrib/pipelines/secadvisor/mod"
+	"github.com/skydive-project/skydive/contrib/exporters/awsflowlogs/mod"
+	"github.com/skydive-project/skydive/contrib/exporters/core"
 )
 
 func main() {
-	core.Main("/etc/skydive/uber.yml")
+	core.Main("/etc/skydive/awsflowlogs.yml")
 }
 
 func init() {
-	core.TransformerHandlers.Register("awsflowlogs", aws.NewTransform, false)
-	core.TransformerHandlers.Register("vpclogs", sa.NewTransform, false)
+	core.TransformerHandlers.Register("awsflowlogs", mod.NewTransform, false)
 }
