+++ contrib/exporters/core/compress.go
