+++ contrib/exporters/core/filter.go
