+++ contrib/exporters/core/pipeline.go
