+++ contrib/exporters/core/s3_client.go
