+++ contrib/exporters/core/store_stdout.go
